refactor: dto 패키지 세분화, 공유하는 dto 분리
dto -> dto.reqeust, dto.response
LineRequest -> LineUpdateRequest
