+++ src/main/java/subway/dto/request/LineRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.request;
 
 public class LineRequest {
