+++ src/main/java/subway/dto/request/LineUpdateRequest.java
@@ -1,0 +1,15 @@
+package subway.dto.request;
+
+public class LineUpdateRequest {
+
+    private String name;
+    private String color;
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+}
