+++ src/main/java/subway/dto/request/SectionAdditionRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.request;
 
 public class SectionAdditionRequest {
     private final Long upStationId;
