+++ src/main/java/subway/ui/StationController.java
@@ -2,8 +2,8 @@ package subway.ui;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.dto.StationRequest;
-import subway.dto.StationResponse;
+import subway.dto.request.StationRequest;
+import subway.dto.response.StationResponse;
 import subway.application.StationService;
 
 import java.net.URI;
